quality_security_platform/backend/requirements.txt
Django==3.2.18
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==22.1
django-environ==0.9.0
#psycopg2-binary==2.9.5
django-redis==5.2.0
redis==4.5.1
celery==5.2.7
drf-yasg==1.21.5
gunicorn==20.1.0
Pillow==9.5.0
django-auth-ldap==5.3.0
pyjwt==2.8.0
requests==2.31.0
django-cas-ng==5.0.0  # 如需 CAS 支持，取消注释
django-constance==4.3.4
# SimpleUI 主题
django-simpleui==2024.3.12